Use prefix pattern for inspect task naming
Renamed all inspect-ai tasks to use inspect_task_ prefix instead of
suffix, improving visual grouping in file listings:

Old pattern (suffix):
- capitalization_inspect_task.py
- twins_inspect_task.py
- general_evaluation_inspect_task.py
- bit_sequences_inspect_task.py
- predictable_or_not_inspect_task.py

New pattern (prefix):
- inspect_task_capitalization.py
- inspect_task_twins.py
- inspect_task_general.py (simplified from general_evaluation)
- inspect_task_bit_sequences.py
- inspect_task_predictable_or_not.py

Benefits:
- All inspect tasks visually cluster together in listings
- Consistent with naming conventions (type comes first)
- Easier to scan glob results
- Cleaner general task name (removed redundant "evaluation")

Updated all references in documentation, READMEs, and task docstrings.
