+++ experiments/synthetic_twins/README.md
@@ -109,20 +109,20 @@ Fine-tuning outputs are saved to:
 After fine-tuning completes, evaluate the model:
 
 ```bash
-# Option 1: Using twins_inspect_task.py (recommended)
-inspect eval experiments/synthetic_twins/twins_inspect_task.py \
+# Option 1: Using inspect_task_twins.py (recommended)
+inspect eval experiments/synthetic_twins/inspect_task_twins.py \
   --model hf/local \
   -M model_path=/path/to/checkpoint \
   -T config_dir=/path/to/epoch_1
 
 # Option 2: Using general evaluation task
-inspect eval experiments/general_evaluation_inspect_task.py \
+inspect eval experiments/inspect_task_general.py \
   --model hf/local \
   -M model_path=/path/to/checkpoint \
   -T config_dir=/path/to/epoch_1
 
 # Evaluate on validation split instead of test
-inspect eval experiments/synthetic_twins/twins_inspect_task.py \
+inspect eval experiments/synthetic_twins/inspect_task_twins.py \
   --model hf/local \
   -M model_path=/path/to/checkpoint \
   -T config_dir=/path/to/epoch_1 \
@@ -167,7 +167,7 @@ experiments/synthetic_twins/
 ├── README.md                    # This file
 ├── preprocess_input_data.py      # Data preprocessing script
 ├── setup_finetune.yaml          # Fine-tuning configuration
-├── twins_inspect_task.py        # Inspect-ai evaluation task
+├── inspect_task_twins.py        # Inspect-ai evaluation task
 ├── finetune.yaml                # Generated by setup_finetune.py
 └── finetune.slurm               # Generated by setup_finetune.py
 ```
