+++ experiments/synthetic_twins/inspect_task_twins.py
@@ -1,18 +1,25 @@
 """
-General Inspect-AI Evaluation Task
+Synthetic Twins Zygosity Prediction - Inspect-AI Evaluation Task
 
-A flexible evaluation task that works with JSON datasets containing input/output pairs.
-Can be used for any experiment in experiments/ directory.
+Evaluates model performance on predicting twin zygosity (monozygotic vs dizygotic)
+from trait values.
 
 Usage:
     # Evaluate fine-tuned model (reads from setup_finetune.yaml)
-    inspect eval experiments/general_evaluation_inspect_task.py --model hf/local -M model_path=/path/to/model -T config_dir=/path/to/epoch_0
+    inspect eval experiments/synthetic_twins/inspect_task_twins.py --model hf/local \\
+        -M model_path=/path/to/model \\
+        -T config_dir=/path/to/epoch_0
 
     # Standalone evaluation with direct dataset path
-    inspect eval experiments/general_evaluation_inspect_task.py --model hf/local -M model_path=/path/to/model -T dataset_path=/path/to/data.json
+    inspect eval experiments/synthetic_twins/inspect_task_twins.py --model hf/local \\
+        -M model_path=/path/to/model \\
+        -T dataset_path=/path/to/twin_zygosity.json
 
-    # With custom parameters
-    inspect eval experiments/general_evaluation_inspect_task.py --model hf/local -M model_path=/path/to/model -T dataset_path=/path/to/data.json -T temperature=0.5 -T system_prompt="You are a helpful assistant"
+    # Evaluate on validation split instead of test
+    inspect eval experiments/synthetic_twins/inspect_task_twins.py --model hf/local \\
+        -M model_path=/path/to/model \\
+        -T config_dir=/path/to/epoch_0 \\
+        -T split=validation
 """
 
 from __future__ import annotations
@@ -22,30 +29,32 @@ from pathlib import Path
 from inspect_ai import Task, task
 from inspect_ai.dataset import hf_dataset, Sample
 from inspect_ai.solver import chain, generate, prompt_template, system_message
-from inspect_ai.scorer import match, includes, accuracy, stderr
+from inspect_ai.scorer import match, includes
 import yaml
 
 
 @task
-def general_eval(
+def twins_zygosity(
     config_dir: Optional[str] = None,
     dataset_path: Optional[str] = None,
     system_prompt: str = "",
-    temperature: float = 0.0,
+    temperature: float = 1e-7,
     split: str = "test",
-    max_tokens: Optional[int] = None
+    max_tokens: int = 10
 ) -> Task:
     """
-    General evaluation task for input/output pair datasets.
+    Evaluation task for twin zygosity prediction.
+
+    Expected output format: "0" (dizygotic) or "1" (monozygotic)
 
     Args:
         config_dir: Path to epoch directory (contains ../setup_finetune.yaml).
                    If provided, reads dataset path and system prompt from config.
         dataset_path: Direct path to dataset JSON file. Used if config_dir not provided.
         system_prompt: System message for the model. Overrides config if both provided.
-        temperature: Generation temperature (default: 0.0 for deterministic output).
-        split: Which data split to use (default: "test").
-        max_tokens: Maximum tokens to generate (default: None, uses model default).
+        temperature: Generation temperature (default: 1e-7 for near-deterministic output).
+        split: Which data split to use - "train", "validation", or "test" (default: "test").
+        max_tokens: Maximum tokens to generate (default: 10, sufficient for binary output).
 
     Returns:
         Task: Configured inspect-ai task
@@ -97,7 +106,7 @@ def general_eval(
             "Must provide either config_dir or dataset_path.\n"
             "Examples:\n"
             "  -T config_dir=/path/to/epoch_0\n"
-            "  -T dataset_path=/path/to/data.json"
+            "  -T dataset_path=/path/to/twin_zygosity.json"
         )
 
     # Handle system prompt if it was parsed as a list (CLI quirk)
@@ -112,12 +121,12 @@ def general_eval(
         )
 
     # Load dataset
-    # Assumes JSON format with structure: {"train": [...], "test": [...]}
+    # Assumes JSON format with structure: {"train": [...], "validation": [...], "test": [...]}
     dataset = hf_dataset(
         path="json",
         data_files=data_path,
         field=split,  # Access the specified split (e.g., "test")
-        split="train",  # Top-level split (confusing but this is how hf_dataset works with nested JSON)
+        split="train",  # Top-level split (HuggingFace quirk - always use "train" here)
         sample_fields=record_to_sample
     )
 
@@ -125,17 +134,17 @@ def general_eval(
     solver_chain = chain(
         system_message(system_prompt),
         prompt_template("{prompt}"),
-        generate({
-            "temperature": temperature,
-            **({"max_tokens": max_tokens} if max_tokens else {})
-        })
+        generate(
+            temperature=temperature,
+            max_tokens=max_tokens
+        )
     )
 
-    # Configure scorers
-    # Using multiple scorers to get different perspectives on accuracy
+    # Configure scorers for binary classification
+    # Using both exact match and includes for robustness
     scorers = [
-        match(location="exact", ignore_case=False),  # Exact match (case-sensitive)
-        includes(ignore_case=False)  # Substring match (case-sensitive)
+        match(location="exact", ignore_case=False),  # Exact match: "0" or "1"
+        includes(ignore_case=False)  # Substring match (for verbose responses)
     ]
 
     return Task(
